Remove endpoint package to enable package default scope for account endpoints and methods.
