+++ src/main/java/com/nexmo/client/applications/DeleteApplicationMethod.java
@@ -19,7 +19,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package com.nexmo.client.applications.endpoints;
+package com.nexmo.client.applications;
 
 import com.nexmo.client.HttpWrapper;
 import com.nexmo.client.NexmoClientException;
