+++ src/main/java/com/nexmo/client/applications/EmbeddedApplicationDetails.java
@@ -19,22 +19,17 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package com.nexmo.client.applications.endpoints;
+package com.nexmo.client.applications;
 
-import com.nexmo.client.NexmoUnexpectedException;
-import com.nexmo.client.applications.ListApplicationsResponse;
-import org.junit.Test;
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import com.fasterxml.jackson.annotation.JsonProperty;
 
-import static org.junit.Assert.fail;
+@JsonIgnoreProperties(ignoreUnknown = true)
+public class EmbeddedApplicationDetails {
+    private ApplicationDetails[] applicationDetails;
 
-public class ListApplicationsResponseTest {
-    @Test
-    public void testJsonError() throws Exception {
-        try {
-            ListApplicationsResponse.fromJson("blarg");
-            fail("Deserializing nonsense JSON should result in a NexmoUnexpectedException");
-        } catch (NexmoUnexpectedException nue) {
-            // This is expected
-        }
+    @JsonProperty("applications")
+    public ApplicationDetails[] getApplicationDetails() {
+        return applicationDetails;
     }
 }
