+++ src/test/java/com/nexmo/client/applications/GetApplicationEndpointTest.java
@@ -19,15 +19,11 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package com.nexmo.client.applications.endpoints;
+package com.nexmo.client.applications;
 
 import com.nexmo.client.HttpConfig;
 import com.nexmo.client.HttpWrapper;
 import com.nexmo.client.TestUtils;
-import com.nexmo.client.applications.ApplicationDetails;
-import com.nexmo.client.applications.ApplicationKeys;
-import com.nexmo.client.applications.UpdateApplicationRequest;
-import com.nexmo.client.applications.WebHook;
 import com.nexmo.client.auth.TokenAuthMethod;
 import org.apache.http.HttpResponse;
 import org.apache.http.client.methods.RequestBuilder;
@@ -38,12 +34,12 @@ import java.util.Map;
 
 import static org.junit.Assert.*;
 
-public class UpdateApplicationMethodTest {
-    private UpdateApplicationMethod endpoint;
+public class GetApplicationEndpointTest {
+    private GetApplicationEndpoint endpoint;
 
     @Before
     public void setUp() throws Exception {
-        this.endpoint = new UpdateApplicationMethod(new HttpWrapper());
+        this.endpoint = new GetApplicationEndpoint(new HttpWrapper());
     }
 
     @Test
@@ -53,55 +49,18 @@ public class UpdateApplicationMethodTest {
     }
 
     @Test
-    public void testMakeRequestWithOptionalParams() throws Exception {
-        UpdateApplicationRequest update = new UpdateApplicationRequest(
-                "app-id",
-                "app name",
-                "https://example.com/answer",
-                "https://example.com/event"
-        );
-        update.setAnswerMethod("PUT");
-        update.setEventMethod("DELETE");
-
-        RequestBuilder builder = this.endpoint.makeRequest(update);
-        assertEquals("PUT", builder.getMethod());
-        assertEquals("https://api.nexmo.com/v1/applications/app-id", builder.build().getURI().toString());
-
-        Map<String, String> params = TestUtils.makeParameterMap(builder.getParameters());
-        assertEquals("app name", params.get("name"));
-        assertEquals("https://example.com/event", params.get("event_url"));
-        assertEquals("https://example.com/answer", params.get("answer_url"));
-        assertEquals("PUT", params.get("answer_method"));
-        assertEquals("DELETE", params.get("event_method"));
-        assertNull(params.get("app_id"));
-    }
-
-    @Test
     public void testMakeRequest() throws Exception {
-        UpdateApplicationRequest update = new UpdateApplicationRequest(
-                "app-id",
-                "app name",
-                "https://example.com/answer",
-                "https://example.com/event"
-        );
-
-        RequestBuilder builder = this.endpoint.makeRequest(update);
-        assertEquals("PUT", builder.getMethod());
+        RequestBuilder builder = this.endpoint.makeRequest("app-id");
+        assertEquals("GET", builder.getMethod());
         assertEquals("https://api.nexmo.com/v1/applications/app-id", builder.build().getURI().toString());
 
         Map<String, String> params = TestUtils.makeParameterMap(builder.getParameters());
-        assertEquals("app name", params.get("name"));
-        assertEquals("https://example.com/event", params.get("event_url"));
-        assertEquals("https://example.com/answer", params.get("answer_url"));
-        assertNull(params.get("answer_method"));
-        assertNull(params.get("event_method"));
-        assertNull(params.get("app_id"));
+        assertEquals(0, params.size());
     }
 
     @Test
     public void testParseResponse() throws Exception {
-        HttpResponse stub = TestUtils.makeJsonHttpResponse(
-                200,
+        HttpResponse stub = TestUtils.makeJsonHttpResponse(200,
                 "{\n" + "  \"id\": \"aaaaaaaa-bbbb-cccc-dddd-0123456789ab\",\n" + "  \"name\": \"My Application\",\n"
                         + "  \"voice\": {\n" + "    \"webhooks\": [\n" + "      {\n"
                         + "        \"endpoint_type\": \"answer_url\",\n"
@@ -110,7 +69,8 @@ public class UpdateApplicationMethodTest {
                         + "        \"endpoint_type\": \"event_url\",\n"
                         + "        \"endpoint\": \"https://example.com/event\",\n"
                         + "        \"http_method\": \"POST\"\n" + "      }\n" + "    ]\n" + "  },\n" + "  \"keys\": {\n"
-                        + "    \"public_key\": \"PUBLIC_KEY\"\n" + "  },\n" + "  \"_links\": {\n" + "    \"self\": {\n"
+                        + "    \"public_key\": \"PUBLIC_KEY\",\n" + "    \"private_key\": \"PRIVATE_KEY\"\n" + "  },\n"
+                        + "  \"_links\": {\n" + "    \"self\": {\n"
                         + "      \"href\": \"/v1/applications/aaaaaaaa-bbbb-cccc-dddd-0123456789ab\"\n" + "    }\n"
                         + "  }\n" + "}"
         );
@@ -135,38 +95,23 @@ public class UpdateApplicationMethodTest {
         ApplicationKeys keys = response.getKeys();
 
         assertEquals("PUBLIC_KEY", keys.getPublicKey());
-        assertNull(keys.getPrivateKey());
+        assertEquals("PRIVATE_KEY", keys.getPrivateKey());
     }
 
     @Test
     public void testDefaultUri() throws Exception {
-        UpdateApplicationRequest request = new UpdateApplicationRequest(
-                "app-id",
-                "app name",
-                "https://example.com/answer",
-                "https://example.com/event"
-        );
-
-        RequestBuilder builder = endpoint.makeRequest(request);
-        assertEquals("PUT", builder.getMethod());
-        assertEquals("https://api.nexmo.com/v1/applications/app-id",
-                builder.build().getURI().toString()
-        );
+        RequestBuilder builder = endpoint.makeRequest("application-id");
+        assertEquals("GET", builder.getMethod());
+        assertEquals("https://api.nexmo.com/v1/applications/application-id", builder.build().getURI().toString());
     }
 
     @Test
     public void testCustomUri() throws Exception {
         HttpWrapper wrapper = new HttpWrapper(new HttpConfig.Builder().baseUri("https://example.com").build());
-        UpdateApplicationMethod method = new UpdateApplicationMethod(wrapper);
-        UpdateApplicationRequest request = new UpdateApplicationRequest(
-                "app-id",
-                "app name",
-                "https://example.com/answer",
-                "https://example.com/event"
-        );
+        GetApplicationEndpoint endpoint = new GetApplicationEndpoint(wrapper);
 
-        RequestBuilder builder = method.makeRequest(request);
-        assertEquals("PUT", builder.getMethod());
-        assertEquals("https://example.com/v1/applications/app-id", builder.build().getURI().toString());
+        RequestBuilder builder = endpoint.makeRequest("application-id");
+        assertEquals("GET", builder.getMethod());
+        assertEquals("https://example.com/v1/applications/application-id", builder.build().getURI().toString());
     }
 }
